Enemy State Machine Partailly
