+++ Assets/Animators New.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5db2d5632acbe6a4cb2eff8088e0acf6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
