+++ Assets/Animators New/Enemy Animators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64a30c95374c999458120ed263a0f66a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
