+++ Assets/Animators New/Enemy Animators/BigEnemy.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac40d3cd648ee7e43a5de48a6a7fc7c6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
