+++ Assets/Animators New/Enemy Animators/Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e98b803298f0e4438e0edfe504a6bf6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
