+++ Assets/EnemyChaseScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28b330939fc55374296c5920e1be13c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
