+++ Assets/Prefabs/Enemy/SmallEnemy.prefab
@@ -446,18 +446,14 @@ GameObject:
   - component: {fileID: 5057332736325729726}
   - component: {fileID: 1054212932288543488}
   - component: {fileID: 1054212932288543490}
-  - component: {fileID: 457103042679402016}
-  - component: {fileID: 3858185729519898145}
   - component: {fileID: 4003223537220944625}
-  - component: {fileID: 9141737296804574616}
-  - component: {fileID: -8467481831254490027}
   m_Layer: 6
   m_Name: SmallEnemy
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &5057332736325729726
 Transform:
   m_ObjectHideFlags: 0
@@ -506,50 +502,6 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 84
   m_CollisionDetection: 0
---- !u!114 &457103042679402016
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5057332736325427294}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 10edcd539a04862439dab6f0ffdd450b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  dustTrailEffects:
-  - {fileID: 1483741861397414457}
-  - {fileID: 7749850405785145632}
---- !u!114 &3858185729519898145
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5057332736325427294}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d963de9cdf5f8b046b0b1cf0ba681022, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  playerTarget: {fileID: 0}
-  enemyStates: 0
-  patrolSpeed: 10
-  chaseSpeed: 12
-  chasingDistance: 20
-  attackDistance: 10
-  currentChaseDistance: 0
-  AttackDistance: 5
-  chaseAfterAtackDistance: 2
-  attackTimer: 0
-  idleAttackTimeGap: 5
-  PatrolMinRadius: 10
-  patrolMaxradius: 20
-  patrolTimer: 0
-  idlePatrolTime: 10
-  tankView: {fileID: 0}
-  enemyShootingBehaviour: {fileID: 9141737296804574616}
 --- !u!195 &4003223537220944625
 NavMeshAgent:
   m_ObjectHideFlags: 0
@@ -557,7 +509,7 @@ NavMeshAgent:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5057332736325427294}
-  m_Enabled: 1
+  m_Enabled: 0
   m_AgentTypeID: 0
   m_Radius: 0.5
   m_Speed: 3.5
@@ -572,36 +524,6 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
---- !u!114 &9141737296804574616
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5057332736325427294}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: df7e4dcbe16e01143b48b2bac78e9674, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  bulletSpawner: {fileID: 2406674086267379875, guid: 8e68e04947fd93d4e83f27dac617d238,
-    type: 3}
-  spawnPoint: {fileID: 1962358364}
---- !u!114 &-8467481831254490027
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5057332736325427294}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4f6339053452bdc48998f0ca005d631d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  explosionPrefab: {fileID: 151086, guid: ee49a9a8678ecf8468eb488e1ac8d42b, type: 3}
-  maxHealth: 100
-  currentHealth: 0
 --- !u!1 &7749850405785145632
 GameObject:
   m_ObjectHideFlags: 0
