+++ Assets/Scenes/Game.unity
@@ -203,7 +203,6 @@ MonoBehaviour:
   view: {fileID: -7696802979863906250, guid: 79c5fa760dedefb48b7ade0ee9b2cfdb, type: 3}
   enemyPrefab:
   - {fileID: -7696802979863906250, guid: 79c5fa760dedefb48b7ade0ee9b2cfdb, type: 3}
-  - {fileID: 457103042679402016, guid: ec0be546c6be05249acb7c899ad95161, type: 3}
   spawnPoints:
   - {fileID: 2046574466}
   - {fileID: 513093858}
@@ -606,7 +605,7 @@ AudioSource:
     type: 2}
   m_audioClip: {fileID: 8300000, guid: 8ce5d020a687c314891341b4126d1de2, type: 3}
   m_PlayOnAwake: 1
-  m_Volume: 0.442
+  m_Volume: 0
   m_Pitch: 1
   Loop: 1
   Mute: 0
