+++ Assets/Scripts/Enemy Tanks MVC/State Machine Behaviour/EnemyChasingState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1638560e55c495b4a88da4ce86e4e119
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
