+++ Assets/Scripts/Enemy Tanks MVC/State Machine Behaviour/EnemyPatrolState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1f05af70c793fa45a10b2d5c50dc964
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
