+++ Assets/Scripts/Player Mvc/New Folder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f25bef554f57b454bb25c55520286404
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
